frontend/requirements.txt
# Core Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# AI/ML Libraries
google-generativeai==0.3.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.3
openpyxl==3.1.2

# HTTP/API
aiohttp==3.8.6
requests==2.31.0
httpx==0.25.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
python-multipart==0.0.6

# Development/Testing
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
mypy==1.5.1